+++ Objects/RoundState.cs
@@ -14,6 +14,7 @@ namespace LethalQuantities.Objects
         public List<SpawnableEnemyWithRarity> daytimeEnemies { get; } = new List<SpawnableEnemyWithRarity>();
         public List<SpawnableEnemyWithRarity> outsideEnemies { get; } = new List<SpawnableEnemyWithRarity>();
 
+        public List<GameObject> modifiedEnemyTypes { get; } = new List<GameObject>();
         public void initialize()
         {
             Plugin.LETHAL_LOGGER.LogInfo("Generating spawnable enemy options");
@@ -31,12 +32,33 @@ namespace LethalQuantities.Objects
             }
         }
 
+        public void OnDestroy()
+        {
+            foreach (GameObject obj in modifiedEnemyTypes)
+            {
+                Destroy(obj);
+            }
+            modifiedEnemyTypes.Clear();
+        }
+
+        public void Update()
+        {
+            foreach (var item in RoundManager.Instance.SpawnedEnemies)
+            {
+                if (item.hideFlags != HideFlags.None)
+                {
+                    item.hideFlags = HideFlags.None;
+                }
+            }
+        }
+
         private void populate(List<SpawnableEnemyWithRarity> enemiesList, List<EnemyTypeConfiguration> configs, bool isOutside = false, bool isDaytimeEnemy = false)
         {
             foreach (var item in configs)
             {
                 int rarity = item.rarity.Value;
-                if (rarity > 0)
+                int maxEnemyCount = item.maxEnemyCount.Value;
+                if (rarity > 0 && maxEnemyCount > 0)
                 {
                     EnemyType type = Instantiate(item.type);
                     type.MaxCount = item.maxEnemyCount.Value;
@@ -64,12 +86,15 @@ namespace LethalQuantities.Objects
             obj.SetActive(false);
 
             GameObject copy = Instantiate(obj, new Vector3(0, -50, 0), Quaternion.identity);
+            modifiedEnemyTypes.Add(copy);
             SceneManager.MoveGameObjectToScene(copy, SceneManager.GetSceneByName("SampleSceneRelay"));
             type.enemyPrefab = copy;
             copy.GetComponent<EnemyAI>().enemyType = type;
             obj.SetActive(isActive);
-            copy.hideFlags = HideFlags.HideInInspector;
+            copy.hideFlags = HideFlags.HideAndDontSave;
             copy.SetActive(isActive);
+
+            DontDestroyOnLoad(copy);
         }
     }
 }